requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

# GPU Acceleration & Deep Learning
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.30.0
accelerate>=0.21.0

# LangChain Ecosystem  
langchain==0.1.0
langchain-community==0.0.10
langchain-experimental>=0.0.50
langchain-core>=0.1.0

# Browser Automation
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Vision & Processing
Pillow==10.1.0
opencv-python>=4.8.0
numpy>=1.24.0

# Database & API
aiosqlite==0.19.0
httpx>=0.25.0
aiofiles==23.2.1
python-multipart==0.0.6

# Utilities
pandas>=2.1.0
structlog>=23.1.0
psutil>=5.9.0
spacy>=3.6.0
nltk>=3.8.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1